requirements.txt
great-expectations==0.17.13
boto3==1.28.39
PyYAML==6.0.1